Move event listener from global to the menu
Instead of updating the menu items with an event listener bound to
global which triggers for any 'restacked' event there is now an event
lister bound to the panel menu button which triggers when the menu is
clicked or touched.

This format of event handling was coppied from the source code for the
panel menu button.