Add key release to events
Window will now also be updated if a key is released. This ensures that
the window will still be updated if accessed by keyboard.

It seems as though the first time clicking the menu after restarting
GNOME Shell is not causing the menu to open. Can't figure out why it
wouldn't work then but will work with every subsequent click.