More accurate signal handling for menu refresh
Menu now only refreshes on 'open-state-changed' signal when menu is
opened. This eliminates pointlessly updating the menu when it is closed.

There is also a new signal handler on the 'restacked' event which will
update the menu if it is currently open. Useful for when a application
finishes loading and creates a window once the menu is already open.