Implement menu sorting
Windows now appear in the menu in the order returned from their
`get-stable-sequence()` method. I have not found out exactly what the
stable sequence is but it seems to reflect the order in which the
windows were opened. This is the sorting order used by the Window List
extension included with GNOME Shell.