Add signal connection for panel updates
Whenever the window menu catches an 'allocate' siganal it checks to see
if the number of items in its panel box has changed. If so it destroys
itself and creates a new window menu. Doing this ensures that if the
menu gets pushed out of its desired position by another extension it
will be put back in the designates spot.

It should be noted that if another extension were to refresh a menu's
position based on a similar set of rules the two menus could probably
push each other into an endless loop. Haven't had a chance to test this.