Refactor with workspace submenus
This is a significant reworking of the menu which uses workspace
submenus when the list workspaces option is enabled. In order to avoid
sporadic resizing of the menu when switching between workspace dropdowns
the menu now has a fixed width. The user option for setting the maximum
window title length to display has been removed as titles are now
automatically cut off when they exceed the menu's set length.